Updating LeftNav naming conventions and fixing scrolling issues
